# docs/swagger.yaml
definitions:
  auth.CreateUserRequest:
    properties:
      email:
        type: string
      fullname:
        maxLength: 64
        minLength: 3
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - fullname
    - password
    - username
    type: object
  auth.LoginRequest:
    properties:
      password:
        maxLength: 32
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  auth.MessageOnlyResponse:
    properties:
      message:
        type: string
    type: object
  auth.SessionResponse:
    properties:
      active_role:
        type: string
      created_at:
        type: string
      expired_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_current:
        type: boolean
      user_agent:
        type: string
    type: object
  utility.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  web.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/utility.APIError'
      success:
        example: false
        type: boolean
    type: object
  web.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  web.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/web.Meta'
      success:
        example: true
        type: boolean
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi user dengan username dan password. Jika berhasil, cookie
        sesi akan di-set pada response.
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: Cookie sesi (X-academic-sesi); HttpOnly; Path=/; Max-Age=86400
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.MessageOnlyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      description: Menghapus sesi user dari database dan menghapus cookie sesi pada
        response.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: Cookie sesi dihapus (X-academic-sesi=; Max-Age=-1)
              type: string
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Logout user
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Menghapus seluruh sesi user yang sedang login dari semua perangkat.
        Berguna saat user mengganti password atau curiga akunnya disalahgunakan.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: Cookie sesi dihapus (X-academic-sesi=; Max-Age=-1)
              type: string
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Logout dari semua perangkat
      tags:
      - auth
  /auth/me:
    get:
      description: Mengembalikan informasi user yang sedang login berdasarkan cookie
        sesi.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Profil user aktif
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Menyimpan user baru
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Buat user
      tags:
      - auth
  /auth/sessions:
    get:
      description: Mengembalikan seluruh sesi aktif (belum expired) milik user yang
        sedang login. Field is_current menandai sesi yang sedang membuat request ini.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/auth.SessionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Daftar sesi aktif user
      tags:
      - auth
swagger: "2.0"